Changed structure and removed button code
Instead of the main GUI class containing the code to render everything
and the logic for everything it now simply contains an ArrayList of
GUIElements each of which will have its render code called each frame.

The GUIElement functionality is fairly basic currently eventually each
element will need to be hooked up to an object or variable in the game
logic branch. Ex: GridElement would have access to the grid variable in
the main game, Ship elements would have access to their associated ships
and the ships would have access to their GUIElement.

Any element which needs to be associated with an object and not a
variable should be initialized in that object's constructor.

Removed the code for buttons and will add them as a new GUIElement next.