Added GUIElementListHandler class
This class is for handling the list of GUIElements and allows for
ordering the list based off of renderPriority, Elements with higher
priority render later each loop and therefore further in the foreground.

It is to be used as a static import to the GUI class, you guys likely
haven't covered those don't worry about it. It lets me reference its
fields and methods without an object or class reference. Ex: If you were
to make Java.util.Math a static import you could just call PI or
max(3,1) instead of Math.PI or Math.max(3,1).